Make command invocation aware of custom commands
The builtin command invocation process used to handle commands defined
in code is not aware of the custom commands defined in the `commands`
module, making it return a "command not found" error, even if there
is a custom command with the same name as the currently invoked command.

This commit makes said native invocation process check if there was any
matched custom command before throwing an error.